third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/state-initially-stuck.html

<!DOCTYPE html>
<title>@container: state(stuck) matching for initial rendering</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  #filler {
    height: 10000px;
  }
  #stuck {
    container-name: initially-stuck;
    container-type: sticky;
    position: sticky;
    bottom: 0;
  }

  @container initially-stuck state(stuck: bottom) {
    span { --stuck: yes }
  }
</style>
<div id="filler"></div>
<div id="stuck">
  <span id="target">My container is stuck</span>
</div>
<script>
  setup(() => assert_implements_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--stuck"), "yes");
  }, "Check that state(stuck: bottom) matches");
</script>
